Move `Handle` and `ObjectId` to dedicated module
